requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
enum34==1.1.10
et-xmlfile==1.1.0
idna==3.7
iniconfig==2.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
pyRofex==0.5.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
